backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newreporter-backend"
version = "0.1.0"
description = "FastAPI backend for NEWREPORTER"
authors = [{ name = "NEWREPORTER Team" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "requests>=2.32.0"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests"
]
pythonpath = [
  "."
]
